move layoutPoints to separate file
